Fixed a bug where the tester does not care if the file ends with a newline
The while loop terminates when the failbit of the istream is set. getline() sets
the failbit of the istream when it does not read any characters. If the istream
contains only the EOF character, getline will not read any characters. This is a
problem when the input ends with a newline since the while loop will not have an
extra iteration compared to input that does not end with a newline.