Removed batch size and strict options
**Batch Size**

Not necessary since the threads will just grab everything in their
subpackage. In the future it would be nice to, instead of
seperating based on packages:
1. Generate the set of all tests
2. Threads access batchSize tests at a time
3. Threads then keep referring back to the test pool for which
    tests need executing

This would eliminate the need for constantly restarting threads,
but would take a bit of a refactor to implement correctly.

**Strict**

I'm tired. I hate writing C++. I'm going to sleep. Here's my PR.